requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.95.1
h11==0.14.0
idna==3.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
PyJWT==2.6.0
python-dotenv==1.0.0
python-ldap==3.4.3
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
